% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{add_graph_summaries}
\alias{add_graph_summaries}
\title{Add Graph Summaries to interaction_model}
\usage{
add_graph_summaries(
  im_input,
  row_key = NULL,
  col_key = NULL,
  tib_with_weights = NULL
)
}
\arguments{
\item{im_input}{An interaction_model object containing an interaction tibble, row universe, and column universe.}

\item{row_key}{The column name in the row universe that corresponds to the node names in the graph.}

\item{col_key}{The column name in the column universe that corresponds to the node names in the graph.}

\item{tib_with_weights}{A tibble containing node weights for personalized PageRank calculation.}
}
\value{
The input interaction_model object with added graph summaries in the row and column universes.
}
\description{
This function computes graph summaries coreness, connected components, and personalized PageRank
for an interaction_model. It adds these summaries to the row and column universes of the input interaction_model.
}
\examples{
im = make_interaction_model(all_packages, ~Package*Imports, parse_text = TRUE, to_lower = FALSE)
row_key = "Package"
col_key = "token"
tib_with_weights = all_packages |> dplyr::mutate(weights = nchar(Package)) |> dplyr::select(Package, weights)
add_graph_summaries(im, row_key, col_key, tib_with_weights)
}
