% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bff.R
\name{bff}
\alias{bff}
\title{Best Feature Function (bff)
Given a pc object from pca, or a rotated version, we wish to interpret the individual dimensions. Often, each unit/row (or context/column) of the original interaction_model will have a some sort of text description.  For example, if each row is an R package, we have the package title and description. Convert these text descriptions into an interaction_model where the units (i.e. the variable before the *) matches either the units or context for the pc.}
\usage{
bff(pcs, im_text, num_best = 10)
}
\arguments{
\item{num_best}{}
}
\description{
This function bff takes both the pc object, and an interaction_model.  It
}
\examples{
im = make_interaction_model(all_packages, ~Package*Imports, parse_text = TRUE, to_lower = FALSE)
pcs = im |> pca(10)
im_text = make_interaction_model(top_packages, ~Package*(Title & Description), parse_text= TRUE)
im_text$column_universe = im_text$column_universe |> dplyr::anti_join(tidytext::stop_words, by = c("token"="word"))
im_text = im_text |> subset_im() |> core()
bff(pcs, im_text, num_best = 4)
}
