% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(fo, tib, is_text = FALSE, make_plot = TRUE)
}
\arguments{
\item{make_plot}{}
}
\description{
This function helps to see that there is "enough data" for the pca to return reliable results. In particular, it examine the degree distribution with both a printout and a plot. Perhaps run this function before running pca.
}
