% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_model.R
\name{.make_interaction_model_completion}
\alias{.make_interaction_model_completion}
\title{Custom completion function for make_interaction_model}
\usage{
.make_interaction_model_completion(.data, formula)
}
\arguments{
\item{.data}{The tibble passed to make_interaction_model}

\item{formula}{The formula passed to make_interaction_model}
}
\description{
Custom completion function for make_interaction_model
}
\keyword{internal}
