% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{extract_interaction}
\alias{extract_interaction}
\title{extract_interaction (internal)}
\usage{
extract_interaction(str)
}
\arguments{
\item{str}{}
}
\value{
a vector of terms inside the interaction
}
\description{
extract_interaction (internal)
}
