% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{get_edge_tib_from_variables}
\alias{get_edge_tib_from_variables}
\title{get_edge_tib_from_variables (internal to make_edge_tib and make_sparse_text_matrix_raw)}
\usage{
get_edge_tib_from_variables(
  tib,
  row_column,
  column_column,
  outcome_column,
  vars = NULL
)
}
\arguments{
\item{outcome_column}{}
}
\description{
get_edge_tib_from_variables (internal to make_edge_tib and make_sparse_text_matrix_raw)
}
