% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{glaplacian}
\alias{glaplacian}
\title{glaplacian}
\usage{
glaplacian(A, regularize = TRUE)
}
\arguments{
\item{regularize}{}
}
\description{
Normalizes and regularizes a sparse adjacency Matrix with deg_row and deg_col as the number of non-zero elements in each row/column
}
