% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{interaction2incomplete}
\alias{interaction2incomplete}
\title{interaction2incomplete}
\usage{
interaction2incomplete(fo, tib)
}
\arguments{
\item{tib}{}
}
\description{
This is like interaction2sparse, but for softImpute.
}
