% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{make_edge_tib}
\alias{make_edge_tib}
\title{make_edge_tib}
\usage{
make_edge_tib(fo, tib)
}
\arguments{
\item{fo}{a formula, like outcome ~ (row_ids & context) * measurement_type.}

\item{tib}{a tibble that contains the variables in the formula. The only exception is that the left-hand-side can be 1 and this does not need to be in tib.}
}
\value{
a list that contains the edge-list tibble, the row_index, and column_index.
}
\description{
make_edge_tib
}
