% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{make_incomplete_matrix_raw}
\alias{make_incomplete_matrix_raw}
\title{make_incomplete_matrix_raw}
\usage{
make_incomplete_matrix_raw(fo, tib)
}
\arguments{
\item{tib}{}
}
\description{
This is like make_sparse_matrix_raw, but for softImpute.
}
