% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_model.R
\name{make_interaction_model}
\alias{make_interaction_model}
\title{make_interaction_model}
\usage{
make_interaction_model(
  .data,
  formula,
  duplicates = "add",
  parse_text = FALSE,
  dropNA = TRUE,
  data_prefix = NULL,
  ...
)
}
\arguments{
\item{.data}{a tibble that contains the variables in the formula. The only exception is that the left-hand-side can be 1 and this does not need to be in .data.}

\item{formula}{a formula, like outcome ~ row_id * (measurement_type & context).}

\item{...}{additional arguments passed to \code{tidytext::unnest_tokens}}
}
\value{
a list with four elements.  First, the interaction_tibble, akin to a sparse matrix in triplet form. Second, row_universe which is akin to the row names of A, but in a tidy form.  Thir, column_universe which is like row_universe. Fourth, some settings.
}
\description{
This generates an \code{interaction_model} object.  If you are comfortable thinking about matrices, you can think of this as a matrix-like-object.
}
\examples{
library(nycflights13)
im = make_interaction_model(flights,~(month & day)*dest)
names(im)
im$row_universe
im$column_universe
im$interaction_tibble
im$settings
# you can extract the sparse Matrix:
A = longpca:::get_Matrix(im,  import_names = TRUE)
str(A)
im = make_interaction_model(all_packages, ~Package*Imports, parse_text = TRUE)
names(im)
im$row_universe
im$column_universe
im$interaction_tibble
im$settings
# with text, there is often a great number of weakly connected words (words that appear once).
# you can remove these words that appear less than 10 times (and documents that have less than 10 words) via:
core(im, core_threshold = 10)
# core retains the k-core of the "largest connected component" in the bipartite graph between rows and columns.

library(dplyr)
# You can provide more than two text columns to make_interaction_model:
im_text = make_interaction_model(top_packages, ~Package*(Title & Description), parse_text= TRUE)
im_text$column_universe |> arrange(desc(n))
# remove stop words by removing them from the column_universe,
#  then use the function subset_im to renumber the columns/rows and remove any lines from interaction_tibble
im_text$column_universe = im_text$column_universe |> anti_join(tidytext::stop_words, by = c("token"="word"))
im_text = im_text |> subset_im()
im_text$column_universe |> arrange(desc(n))
}
