% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{make_sparse_matrix_raw}
\alias{make_sparse_matrix_raw}
\title{make_sparse_matrix_raw}
\usage{
make_sparse_matrix_raw(fo, tib, dropNA = TRUE)
}
\arguments{
\item{fo}{a formula, like outcome ~ (row_ids & context) * measurement_type.}

\item{tib}{a tibble that contains the variables in the formula. The only exception is that the left-hand-side can be 1 and this does not need to be in tib.}

\item{dropNA}{recommended.  This drops rows of tib if there are any NA's among the essential variables.}
}
\value{
a list with three elements.  First, the sparse Matrix A. Second, row_universe which is akin to the row names of A, but in a tidy form.  Thir, column_universe which is like row_universe.
}
\description{
make_sparse_matrix_raw
}
