% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_matrix.R
\name{make_sparse_text_matrix_raw}
\alias{make_sparse_text_matrix_raw}
\title{make_sparse_text_matrix_raw}
\usage{
make_sparse_text_matrix_raw(fo, tib, dropNA = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
make_sparse_text_matrix_raw
}
