% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_model.R
\name{old_make_interaction_model}
\alias{old_make_interaction_model}
\title{old_make_interaction_model}
\usage{
old_make_interaction_model(
  fo,
  tib,
  duplicates = "add",
  parse_text = FALSE,
  dropNA = TRUE,
  data_prefix = NULL,
  ...
)
}
\arguments{
\item{fo}{a formula, like outcome ~ (row_nums & context) * measurement_type.}

\item{tib}{a tibble that contains the variables in the formula. The only exception is that the left-hand-side can be 1 and this does not need to be in tib.}

\item{parse_text}{if set to TRUE, then the right side of * (i.e. the measurement_type) will be parsed as a sequence of tokens via tidytext::unnest_tokens.  Additional arguments in ... will be passed to unnest_tokens.  For example, adding to_lower = FALSE will ensure case is kept. Additionally, you could set token to be something other than words (ngrams or skip_ngrams and then additionally specify n = ).  See unnest_token for more arguments.}

\item{dropNA}{recommended.  This drops rows of tib if there are any NA's among the essential variables.}
}
\value{
a list with four elements.  First, the interaction_tibble, akin to a sparse matrix in triplet form. Second, row_universe which is akin to the row names of A, but in a tidy form.  Thir, column_universe which is like row_universe. Fourth, some settings.
}
\description{
This generates an \code{interaction_model} object.  If you are comfortable thinking about matrices, you can think of this as a matrix-like-object.
}
\examples{
library(nycflights13)
im = old_make_interaction_model(~(month & day)*dest, flights)
names(im)
im$row_universe
im$column_universe
im$interaction_tibble
im$settings
# you can extract the sparse Matrix:
A = longpca:::get_Matrix(im,  import_names = TRUE)
str(A)
im = old_make_interaction_model(~Package*Imports, all_packages, parse_text = TRUE)
names(im)
im$row_universe
im$column_universe
im$interaction_tibble
im$settings
}
