% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_na}
\alias{pca_na}
\title{pca_na}
\usage{
pca_na(im, k)
}
\arguments{
\item{k}{}
}
\description{
This performs pca on an interaction_model, where if a (row_id,col_id) are not present in the data, then it is presumed NA and imputed.  This contrasts to the other \code{pca} function which imputes the missing values to zero.
}
