% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_sum}
\alias{pca_sum}
\title{pca_sum}
\usage{
pca_sum(fo, tib, k)
}
\arguments{
\item{fo}{a formula, like  outcome ~ (row_ids & context) * measurement_type.}

\item{tib}{a tibble that contains the variables in the formula. The only exception is that the left-hand-side can be 1 and this does not need to be in tib.}

\item{k}{number of dimensions for the pca}
}
\value{
the output is a list of: (1) row_features for whichever term is first on the right hand side of the formula, (2) column_features for whichever term is second on the right hand side of the formula, (3) middle_B which is a tibble corresponding to a diagonal matrix in sparse triplet form, (4) settings which is a list of details.
}
\description{
This performs pca on a sparse Matrix specified by the formula and the tibble.
}
