% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_dim.R
\name{pick_dim}
\alias{pick_dim}
\title{pick_dim}
\usage{
pick_dim(im, dimMax = 20, num_bootstraps = 2)
}
\arguments{
\item{num_bootstraps}{}
}
\description{
This computes the Z-scores from cross-validated eigenvalues
}
\examples{
library(nycflights13)
im = make_interaction_model(flights,~(month&day)*dest)
cveig = pick_dim(im, dimMax = 7)
plot(cveig)
cveig
}
