% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_dim.R
\name{pick_dim_text}
\alias{pick_dim_text}
\title{pick_dim_text}
\usage{
pick_dim_text(fo, tib, dimMax = 20, num_bootstraps = 2, ...)
}
\arguments{
\item{num_bootstraps}{}
}
\description{
This computes the Z-scores from cross-validated eigenvalues for a bag-of-words matrix.
}
\details{
The arguments ... get handed to unnest_tokens. See Details of pca_text for mmore.
}
