% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_model.R
\name{print.interaction_model}
\alias{print.interaction_model}
\title{print interaction model}
\usage{
\method{print}{interaction_model}(im)
}
\arguments{
\item{im}{}
}
\description{
print interaction model
}
