% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{remove_L_normalization}
\alias{remove_L_normalization}
\title{remove_L_normalization (internal to pca_mean/pca_average)}
\usage{
remove_L_normalization(s_svd, A, orthogonalize = FALSE)
}
\arguments{
\item{orthogonalize}{}
}
\description{
remove_L_normalization (internal to pca_mean/pca_average)
}
