% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{update_lhs_to_1}
\alias{update_lhs_to_1}
\title{update_lhs_to_1 (internal to pca_count)}
\usage{
update_lhs_to_1(formula)
}
\arguments{
\item{formula}{}
}
\description{
update_lhs_to_1 (internal to pca_count)
}
