% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_dim.R
\name{pick_dim}
\alias{pick_dim}
\title{pick_dim}
\usage{
pick_dim(fo, tib, dimMax = 20, num_bootstraps = 2)
}
\arguments{
\item{num_bootstraps}{}
}
\description{
This computes the Z-scores from cross-validated eigenvalues
}
