% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{subset_im}
\alias{subset_im}
\title{subset_im
If you've removed some rows of row_universe or column_universe, then you want to redefine a bunch of things... you need new row/col numbering and you want to remove rows from interaction_tibble.  this does that.}
\usage{
subset_im(im_input)
}
\arguments{
\item{im_input}{}
}
\description{
subset_im
If you've removed some rows of row_universe or column_universe, then you want to redefine a bunch of things... you need new row/col numbering and you want to remove rows from interaction_tibble.  this does that.
}
